Convert fletching to use typescript
