+++ src/main/javascript/modules/skill/fletching/index.ts
@@ -1,0 +1,3 @@
+import './string-bow';
+import './feather-arrow';
+import './fletch-log';
