+++ src/main/javascript/modules/skill/index.ts
@@ -1,5 +1,6 @@
 import './agility';
 import './fishing';
+import './fletching';
 import './mining';
 import './thieving';
 import './woodcutting';
